// resources/js/api/generated/models/component.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface Component {
  /** @minimum 0 */
  bytes_added: number;
  /** @minimum 0 */
  bytes_processed: number;
  duration_secs: number;
  error?: string;
  excludes?: string[];
  /** @minimum 0 */
  files_changed: number;
  /** @minimum 0 */
  files_new: number;
  /** @minimum 0 */
  files_unmodified: number;
  id: string;
  kind: string;
  service?: string;
  snapshot_id?: string;
  source_path?: string;
  target?: string;
  volume_name?: string;
}