// resources/js/api/generated/models/runSummary.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface RunSummary {
  /** @minimum 0 */
  added_bytes: number;
  component_count: number;
  components_with_errors: number;
  /** @nullable */
  finished_at?: string | null;
  id: string;
  label?: string;
  /** @minimum 0 */
  repo_size_bytes?: number;
  /** @minimum 0 */
  size_bytes: number;
  stack_name: string;
  started_at: string;
  status: string;
  stop_mode?: string;
  /** @nullable */
  verified?: boolean | null;
}